fix context crossing interrupt handling and 64-bit accounting
Move interrupt disable from guest C caller to assembly (between CR3 and
lidt). The switch call invokes a complete monitor-loop iteration; cli in
the caller was held across everything, causing freezes on certain exit
paths. The five-instruction critical section is now properly protected
in the switch.

Also: fix 64-bit overflow in hostmem fields, remove unused reversed pfn
mapping, add terminal keys. Verified on M92p: KVM switch and ctrl-C.

+++ src/colinux/common/ioctl.h
@@ -115,8 +115,8 @@ typedef struct {
 
 /* interface for CO_MANAGER_IOCTL_INFO: */
 typedef struct {
-	unsigned long hostmem_usage_limit;
-	unsigned long hostmem_used;
+	unsigned long long hostmem_usage_limit;
+	unsigned long long hostmem_used;
 } co_manager_ioctl_info_t;
 
 #define CO_MANAGER_ATTACH_MAX_MODULES 0x10
